Wire up fields for MailChimp stuff on the tab form.
The tab form now has a dropdown that allows the user to specify what
external service they would like to use, and two fields to specify
details on their MailChimp account. Currently, you have to do a bit of
hunting to find the information you need to populate this form. When I
implement the manager design, I'll make it more user friendly.