Disallow replay attacks on Stripe notifications.
This isn't quite as ambitious as I wanted to be for this pass, but it'll
do for now. We don't have a way to get access to the Event collection.
Spoofing is technically possible at the moment if you have valid ID's
and such, but we're not a high priority target. So, we'll worry about
that later.

closes gh-59