Implement bootstrapping of database version.
This is the groundwork for automated migrations. We introduce the
DatabaseMeta MongoDocument that allows us to store key/values about the
current Mongo database. We then ensure, on boot, that some database
version is set. If not, we set it. This also results in the moving of
migrations into the resources folder so they appear in the JAR.