Implement the iframe parameters for ServiceWrapper.
When present in the service wrapper the tab is hooked up to, the success
response from a submit will include details about a URI that the client
should load in an hidden iframe on submit. This is ideal for service
wrappers where we don't integrate directly with the upstream service for
one reason or another.