Fix issues with billing failures giving an extra month of subscription.
We were improperly checking for the "unpaid" status on subscriptions,
when our account settings don't behave that way. When charges fail, we
cancel the subscription outright. Now we're properly checking the
cancel_at_period_end flag, and responding correctly.

We also no longer nuke the user's card information. There's no
information in the event that I can find to tell us this is actually a
case of us not being paid, so I don't want to nuke card information in
the event we just canceled it from stripe.